scripts

+++ removeExtensions.sh
@@ -1,3 +1,9 @@
 cd ~/.vscode/extensions
-rm -rf *
-echo "All extensions from vscode have been removed. You will have to reset vscode"
+echo "Are you sure that you want to remove all extensions"
+read -p "YES OR NO: " YesOrNo
+if [[ $YesOrNo == "YES" ]] then
+    rm -rf *
+    echo "All extensions from vscode have been removed. You will have to reset vscode"
+else
+    exit
+fi
